COMP:Removed useless/old-style cast.

+++ VTK/Widgets/vtkAngleRepresentation3D.cxx
@@ -273,7 +273,7 @@ void vtkAngleRepresentation3D::BuildRepresentation()
 
         vtkPoints *points = this->ArcSource->GetOutput()->GetPoints();
         const int npoints = points->GetNumberOfPoints();
-        points->GetPoint( (int)(npoints / 2), this->TextPosition );
+        points->GetPoint(npoints/2, this->TextPosition );
 
         char string[512];
         sprintf(string, this->LabelFormat,
